Prototype dataset explorer and preview (#4)
Show a list of data sources with some information and a preview button that shows the data in a table. Also allows searching/filtering datasets. Uses material-ui module for styling, react-table for table logic, and react-select for dataset filtering. Note that this is using direct API calls to a few example data sources, this will be replaced with calls to the backend once we have that set up.